requirements.txt
discord.py
Pillow
requests
pynacl
youtube_dl
pony
psycopg2
agithub
pytumblr
-e git+https://github.com/HarkonenBade/ao3.git@multi-author#egg=ao3
html2text
